scripts/cloud_functions/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-storage==2.0.0
